launchCohortExplorer {LegendT2dmCohortExplorer}	R Documentation

Launch the Cohort/Diagnostics Explorer Shiny app

Description

Launch the Cohort/Diagnostics Explorer Shiny app

Usage

launchCohortExplorer(
  dataFolder = "data",
  dataFile = "PreMerged.RData",
  connectionDetails = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchemas = resultsDatabaseSchema,
  aboutText = NULL,
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE
)

Arguments

dataFolder
A folder where the premerged file is stored. Use the preMergeDiagnosticsFiles function to generate this file.

dataFile
(Optional) The name of the .RData file with results. It is commonly known as the Premerged file.

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package, specifying how to connect to the server where the study results results have been uploaded using the uploadResults function.

resultsDatabaseSchema
The schema on the database server where the results have been uploaded.

vocabularyDatabaseSchemas
(optional) A list of one or more schemas on the database server where the vocabulary tables are located. The default value is the value of the resultsDatabaseSchema. We can provide a list of vocabulary schema that might represent different versions of the OMOP vocabulary tables. It allows us to compare the impact of vocabulary changes on Diagnostics.

aboutText
Text (using HTML markup) that will be displayed in an About tab in the Shiny app. If not provided, no About tab will be shown.

runOverNetwork
(optional) Do you want the app to run over your network?

port
(optional) Only used if runOverNetwork = TRUE.

launch.browser
Should the app be launched in your default browser, or in a Shiny window. Note: copying to clipboard will not work in a Shiny window.

appDir
Top directory of Shiny app

Details

Launches a Shiny app that allows the user to explore the evidence

[Package LegendT2dmCohortExplorer version 0.1.0 Index]